// extrabeam-backend/src/types/database.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[];

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: '13.0.4';
  };
  public: {
    Tables: {
// ...
      unavailabilities: {
        Row: {
          created_at: string;
          end_time: string;
          entreprise_id: number;
          exceptions: Json | null;
          id: number;
          recurrence_end: string | null;
          recurrence_type: string;
          start_date: string;
          start_time: string;
          title: string | null;
          updated_at: string;
          weekday: number | null;
        };
        Insert: {
          created_at?: string;
          end_time: string;
          entreprise_id: number;
          exceptions?: Json | null;
          id?: never;
          recurrence_end?: string | null;
          recurrence_type?: string;
          start_date: string;
          start_time: string;
          title?: string | null;
          updated_at?: string;
          weekday?: number | null;
        };
        Update: {
          created_at?: string;
          end_time?: string;
          entreprise_id?: number;
          exceptions?: Json | null;
          id?: never;
          recurrence_end?: string | null;
          recurrence_type?: string;
          start_date?: string;
          start_time?: string;
          title?: string | null;
          updated_at?: string;
          weekday?: number | null;
        };
        Relationships: [
          {
            foreignKeyName: 'unavailabilities_company_id_fkey';
            columns: ['entreprise_id'];
            isOneToOne: false;
            referencedRelation: 'entreprise';
            referencedColumns: ['id'];
          },
        ];
      };
    };
// ...
  };
};